attempted to fix rules

+++ syntax_analysis/sketches/BisonExample/includes/stack_operations.hpp
@@ -1,2 +1,2 @@
-uint64_t get_variable(std::map<std::string, uint64_t>, std::string);
+double get_variable(std::map<std::string, double>, std::string);
 void peetr();
