+++ syntax_analysis/sketches/BisonExample/src/stack_operations.cpp
@@ -5,13 +5,13 @@
 
 #include "../includes/stack_operations.hpp"
 
-uint64_t get_variable(std::map<std::string, uint64_t> container, std::string term){
+double get_variable(std::map<std::string, double> container, std::string term){
   for (auto it = container.begin(); it != container.end(); ++it){
     if (it->first == term){
       return it->second;
     }
   }
-  return std::numeric_limits<uint64_t>::infinity();
+  return std::numeric_limits<double>::infinity();
 }
 
 void peetr(){
